refactor: update sql (add cols, update type)

+++ sql/user_goals_saving.sql
@@ -10,7 +10,7 @@ CREATE TABLE goal_setting
     goal_type           VARCHAR(20),
     priority_level      INTEGER                   CHECK (priority_level IN (1, 2, 3)),
     weight              NUMERIC(5,2),
-    status              INTEGER                  DEFAULT 0,
+    status              TINYINT                  DEFAULT 0,
     eta_lock            BOOLEAN                   DEFAULT FALSE,
     initial_target_date DATE,
     sent_money          BOOLEAN                   DEFAULT FALSE,
